# docs/api/producer-portal.openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Insuryn producer portal",
        "version": "1.0.0",
        "description": "Read-only producer portal (my customers and policies, renewals due, collections to deposit, statements, licence, targets) plus recording cash collections. Every request names the tenant (X-Tenant header or tenant subdomain); tokens only work in the tenant that issued them."
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
        "/api/portal/collections": {
            "post": {
                "operationId": "recordCollection",
                "summary": "Record cash I collected for one of my policies",
                "tags": [
                    "Producer portal"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Tenant"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CollectionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CollectionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "x-token-ability": "portal:collect"
            }
        },
        "/api/portal/collections-to-deposit": {
            "get": {
                "operationId": "collectionsToDeposit",
                "summary": "Cash I collected and have not deposited yet",
                "tags": [
                    "Producer portal"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Tenant"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CollectionsToDepositResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "x-token-ability": "portal:read"
            }
        },
        "/api/portal/customers": {
            "get": {
                "operationId": "customers",
                "summary": "My customers: policyholders of my policies",
                "tags": [
                    "Producer portal"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Tenant"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerList"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "x-token-ability": "portal:read"
            }
        },
        "/api/portal/licence": {
            "get": {
                "operationId": "licence",
                "summary": "My licence status",
                "tags": [
                    "Producer portal"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Tenant"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LicenceResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "x-token-ability": "portal:read"
            }
        },
...
        "/api/portal/policies": {
            "get": {
                "operationId": "policies",
                "summary": "My policies",
                "tags": [
                    "Producer portal"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Tenant"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "description": "issued, active, lapsed, cancelled, expired, renewed"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PolicyList"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "x-token-ability": "portal:read"
            }
        },
...
        "/api/portal/renewals-due": {
            "get": {
                "operationId": "renewalsDue",
                "summary": "My policies due for renewal",
                "tags": [
                    "Producer portal"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Tenant"
                    },
                    {
                        "name": "within_days",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Days ahead, default 60"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PolicyList"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "x-token-ability": "portal:read"
            }
        },
        "/api/portal/statements": {
            "get": {
                "operationId": "statements",
                "summary": "My commission statements",
                "tags": [
                    "Producer portal"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Tenant"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatementList"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "x-token-ability": "portal:read"
            }
        },
        "/api/portal/statements/{statement}": {
            "get": {
                "operationId": "statement",
                "summary": "One of my statements, with its entries",
                "tags": [
                    "Producer portal"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Tenant"
                    },
                    {
                        "name": "statement",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatementDetailResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "x-token-ability": "portal:read"
            }
        },
        "/api/portal/targets": {
            "get": {
                "operationId": "targets",
                "summary": "My targets and achievement for a period",
                "tags": [
                    "Producer portal"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Tenant"
                    },
                    {
                        "name": "period_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "description": "monthly, quarterly or annual (default monthly)"
                    },
                    {
                        "name": "period_start",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "description": "First day of the period (default: this month)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TargetList"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Refused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "x-token-ability": "portal:read"
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "description": "Sanctum personal access token from POST /api/portal/tokens"
            }
        },
        "parameters": {
            "Tenant": {
                "name": "X-Tenant",
                "in": "header",
                "required": false,
                "schema": {
                    "type": "string"
                },
                "description": "Tenant id or slug, when not given by the subdomain"
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "required": [
                    "message",
                    "reason"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "reason": {
                        "type": "string"
                    }
                }
            },
...
            "CollectionRequest": {
                "type": "object",
                "required": [
                    "installment_id",
                    "amount_minor",
                    "value_date"
                ],
                "properties": {
                    "installment_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "amount_minor": {
                        "type": "integer",
                        "description": "Minor units"
                    },
                    "value_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "reference": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            },
...
            "CollectionResponse": {
                "type": "object",
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Collection"
                    }
                }
            },
...
        }
    },
    "security": [
        {
            "bearer": []
        }
    ]
}